Decouple Map and SvgMap
